Post initial C code to master branch
This code is not commented well enough in its current form. I will be updating it piece by piece in time, which it will take; as a reminder to all be sure to comment your code as you go so you do not have to deal with this. Expect to see several commits as these files are updated and as certain parts are grouped together, such as those that require text files to properly operate.